cookiecutter-repo/requirements.txt
click==8.1.3
hydra-core==1.3.1
matplotlib==3.6.2
numpy==1.24.1
omegaconf==2.3.0
pytest==7.2.0
python-dotenv==0.21.0
pytorch_lightning==1.8.6
scikit_learn==1.2.0
setuptools==65.5.0
torch==1.13.1
torchvision==0.14.1
